update file directory -- move consistent hashing under util folder and add union functions to util folder
