+++ globalscheduler/controllers/scheduler/scheduler_controller.go
@@ -44,7 +44,8 @@ import (
 	schedulerlisters "k8s.io/kubernetes/globalscheduler/pkg/apis/scheduler/client/listers/scheduler/v1"
 	schedulercrdv1 "k8s.io/kubernetes/globalscheduler/pkg/apis/scheduler/v1"
 
-	"k8s.io/kubernetes/globalscheduler/controllers/scheduler/consistenthashing"
+	"k8s.io/kubernetes/globalscheduler/controllers/util/consistenthashing"
+	"k8s.io/kubernetes/globalscheduler/controllers/util/union"
 )
 
 const (
@@ -57,7 +58,7 @@ type SchedulerController struct {
 	// kubeclientset is a standard kubernetes clientset
 	kubeclientset   *kubernetes.Clientset
 	schedulerclient *schedulerclient.SchedulerClient
-	clusterclient *clusterclient.ClusterClient
+	clusterclient   *clusterclient.ClusterClient
 
 	schedulerInformer schedulerlisters.SchedulerLister
 	clusterInformer   clusterlisters.ClusterLister
@@ -114,9 +115,9 @@ func NewSchedulerController(
 	})
 
 	clusterInformer.Informer().AddEventHandler(cache.ResourceEventHandlerFuncs{
-		AddFunc:    controller.addClusterToScheduler,
+		AddFunc: controller.addClusterToScheduler,
 		//UpdateFunc: controller.updateClusterFromScheduler,
-		//DeleteFunc: controller.deleteClusterFromScheduler,
+		DeleteFunc: controller.deleteClusterFromScheduler,
 	})
 
 	return controller
@@ -274,7 +275,7 @@ func (sc *SchedulerController) syncHandler(key *KeyWithEventType) error {
 		schedulerCopy.Spec.Cluster = append(schedulerCopy.Spec.Cluster, cluster)
 
 		// Union
-		schedulerCopy = unionUpdate(schedulerCopy, cluster)
+		schedulerCopy = union.UpdateUnion(schedulerCopy, cluster)
 
 		_, err = sc.schedulerclient.Update(schedulerCopy)
 		if err != nil {
@@ -283,6 +284,28 @@ func (sc *SchedulerController) syncHandler(key *KeyWithEventType) error {
 		}
 
 		sc.recorder.Event(scheduler, corev1.EventTypeNormal, SuccessSynced, MessageResourceSynced)
+
+	case EventTypeDeleteCluster:
+		cluster, err := sc.clusterInformer.Clusters(namespace).Get(name)
+		if err != nil {
+			// The cluster resource may no longer exist, in which case we stop
+			// processing.
+			if errors.IsNotFound(err) {
+				runtime.HandleError(fmt.Errorf("failed to list cluster by: %s/%s", namespace, name))
+				return nil
+			}
+			return err
+		}
+
+		scheduler, err := sc.schedulerclient.Get(cluster.Spec.HomeScheduler, metav1.GetOptions{})
+		if err != nil {
+			klog.Infof("Error getting scheduler object")
+			return err
+		}
+
+		schedulerCopy := scheduler.DeepCopy()
+		schedulerCopy.Spec.Cluster = union.RemoveCluster(schedulerCopy.Spec.Cluster, cluster)
+
 	}
 
 	return nil
@@ -301,10 +324,10 @@ func (sc *SchedulerController) addClusterToScheduler(clusterObj interface{}) {
 //	fmt.Println("Watching Cluster Update")
 //}
 //
-//func (sc *SchedulerController) deleteClusterFromScheduler(clusterObj interface{}) {
-//	cluster := clusterObj.(*clustercrdv1.Cluster)
-//	sc.enqueue(cluster, EventTypeDeleteCluster)
-//}
+func (sc *SchedulerController) deleteClusterFromScheduler(clusterObj interface{}) {
+	cluster := clusterObj.(*clustercrdv1.Cluster)
+	sc.enqueue(cluster, EventTypeDeleteCluster)
+}
 
 func (sc *SchedulerController) addScheduler(schedulerObj interface{}) {
 	sc.enqueue(schedulerObj, EventTypeCreateScheduler)
@@ -346,184 +369,4 @@ func (sc *SchedulerController) enqueue(obj interface{}, eventType EventType) {
 	}
 	keyWithEventType := NewKeyWithEventType(eventType, key)
 	sc.workqueue.AddRateLimited(keyWithEventType)
-}
-
-func unionUpdate(schedulerCopy *schedulercrdv1.Scheduler, cluster *clustercrdv1.Cluster) *schedulercrdv1.Scheduler {
-	union := schedulerCopy.Spec.Union
-
-	// IpAddress Union
-	union.IpAddress = unionIpAddress(union.IpAddress, cluster.Spec.IpAddress)
-
-	// GeoLocation Union
-	union.GeoLocation = unionGeoLocation(union.GeoLocation, cluster.Spec.GeoLocation)
-
-	// Region Union
-	union.Region = unionRegion(union.Region, cluster.Spec.Region)
-
-	// Operator Union
-	union.Operator = unionOperator(union.Operator, cluster.Spec.Operator)
-
-	// Flavors Union
-	union.Flavors = unionFlavors(union.Flavors, cluster.Spec.Flavors)
-
-	// Storage Union
-	union.Storage = unionStorage(union.Storage, cluster.Spec.Storage)
-
-	// EipCapacity Union
-	union.EipCapacity = unionEipCapacity(union.EipCapacity, cluster.Spec.EipCapacity)
-
-	// CPUCapacity Union
-	union.CPUCapacity = unionCPUCapacity(union.CPUCapacity, cluster.Spec.CPUCapacity)
-
-	// MemCapacity Union
-	union.MemCapacity = unionMemCapacity(union.MemCapacity, cluster.Spec.MemCapacity)
-
-	// ServerPrice Union
-	union.ServerPrice = unionServerPrice(union.ServerPrice, cluster.Spec.ServerPrice)
-
-	schedulerCopy.Spec.Union = union
-	return schedulerCopy
-}
-
-func unionStorage(unionStorage []*clustercrdv1.StorageSpec, storage []clustercrdv1.StorageSpec) []*clustercrdv1.StorageSpec {
-	m := make(map[*clustercrdv1.StorageSpec]int)
-	for _, v := range unionStorage {
-		m[v]++
-	}
-
-	for _, v := range storage {
-		times, _ := m[&v]
-		if times == 0 {
-			unionStorage = append(unionStorage, &v)
-		}
-	}
-	return unionStorage
-}
-
-func unionFlavors(unionFlavors []*clustercrdv1.FlavorInfo, flavors []clustercrdv1.FlavorInfo) []*clustercrdv1.FlavorInfo {
-	m := make(map[*clustercrdv1.FlavorInfo]int)
-	for _, v := range unionFlavors {
-		m[v]++
-	}
-
-	for _, v := range flavors {
-		times, _ := m[&v]
-		if times == 0 {
-			unionFlavors = append(unionFlavors, &v)
-		}
-	}
-	return unionFlavors
-
-}
-
-func unionServerPrice(unionPrice []int64, price int64) []int64 {
-	m := make(map[int64]int)
-	for _, v := range unionPrice {
-		m[v]++
-	}
-
-	times, _ := m[price]
-	if times == 0 {
-		unionPrice = append(unionPrice, price)
-	}
-
-	return unionPrice
-}
-
-func unionMemCapacity(unionMemCapacity []int64, memCapacity int64) []int64 {
-	m := make(map[int64]int)
-	for _, v := range unionMemCapacity {
-		m[v]++
-	}
-
-	times, _ := m[memCapacity]
-	if times == 0 {
-		unionMemCapacity = append(unionMemCapacity, memCapacity)
-	}
-
-	return unionMemCapacity
-}
-
-func unionCPUCapacity(unionCPUCapacity []int64, cpuCapacity int64) []int64 {
-	m := make(map[int64]int)
-	for _, v := range unionCPUCapacity {
-		m[v]++
-	}
-
-	times, _ := m[cpuCapacity]
-	if times == 0 {
-		unionCPUCapacity = append(unionCPUCapacity, cpuCapacity)
-	}
-
-	return unionCPUCapacity
-}
-
-func unionEipCapacity(unionEipCapacity []int64, eipCapacity int64) []int64 {
-	m := make(map[int64]int)
-	for _, v := range unionEipCapacity {
-		m[v]++
-	}
-
-	times, _ := m[eipCapacity]
-	if times == 0 {
-		unionEipCapacity = append(unionEipCapacity, eipCapacity)
-	}
-
-	return unionEipCapacity
-}
-
-func unionOperator(unionOperator []*clustercrdv1.OperatorInfo, operator clustercrdv1.OperatorInfo) []*clustercrdv1.OperatorInfo {
-	m := make(map[*clustercrdv1.OperatorInfo]int)
-	for _, v := range unionOperator {
-		m[v]++
-	}
-
-	times, _ := m[&operator]
-	if times == 0 {
-		unionOperator = append(unionOperator, &operator)
-	}
-
-	return unionOperator
-}
-
-func unionRegion(unionRegion []*clustercrdv1.RegionInfo, region clustercrdv1.RegionInfo) []*clustercrdv1.RegionInfo {
-	m := make(map[*clustercrdv1.RegionInfo]int)
-	for _, v := range unionRegion {
-		m[v]++
-	}
-
-	times, _ := m[&region]
-	if times == 0 {
-		unionRegion = append(unionRegion, &region)
-	}
-
-	return unionRegion
-}
-
-func unionGeoLocation(unionGeoLocation []*clustercrdv1.GeolocationInfo, geoLocation clustercrdv1.GeolocationInfo) []*clustercrdv1.GeolocationInfo {
-	m := make(map[*clustercrdv1.GeolocationInfo]int)
-	for _, v := range unionGeoLocation {
-		m[v]++
-	}
-
-	times, _ := m[&geoLocation]
-	if times == 0 {
-		unionGeoLocation = append(unionGeoLocation, &geoLocation)
-	}
-
-	return unionGeoLocation
-}
-
-func unionIpAddress(unionIp []string, ip string) []string {
-	m := make(map[string]int)
-	for _, v := range unionIp {
-		m[v]++
-	}
-
-	times, _ := m[ip]
-	if times == 0 {
-		unionIp = append(unionIp, ip)
-	}
-
-	return unionIp
 }
